pyproject.toml
[project]
name = "chatbot01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents[litellm]>=0.0.14",
    "python-dotenv>=1.1.0",
]
